fix(make-reservation: fix method with the properties of the class Lodging
