+++ src/main/java/com/bookstay/DayResort.java
@@ -1,39 +1,48 @@
-package com.bookstay.models;
+package com.bookstay;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.List;
 
-public class FarmStay extends Lodging{
-    private int maxCapacity;
-    private double pricePerNight;
+public class DayResort extends Lodging {
+    private List<String> activities;
+    private List<String> meals;
+    private double pricePerPerson;
 
-    public FarmStay(String name, String city, double rating, String description, int maxCapacity, double price) {
-        super(name, city, "Finca", rating, description);
-        this.maxCapacity = maxCapacity;
-        this.pricePerNight = price;
+    public DayResort(String name, String city, double rating, String description, double price) {
+        super(name, city, "Día de sol", rating, description);
+        this.activities = new ArrayList<>();
+        this.meals = new ArrayList<>();
+        this.pricePerPerson = price;
+    }
+
+    public void addActivity(String activity){
+        activities.add(activity);
+    }
+
+    public void addMeal(String meal){
+        meals.add(meal);
     }
 
     @Override
     public double calculatePrice(int adults, int children, int days) {
-        int totalPeople = adults + children;
-
-        if (totalPeople > maxCapacity) {
-            throw new IllegalArgumentException("La capacidad máxima de la finca es de " + maxCapacity + " personas.");
+        if (days != 1) {
+            throw new IllegalArgumentException("Un 'Día de Sol' solo puede ser reservado por 1 día.");
         }
 
-        return pricePerNight * days;
+        int totalGuests = adults + children;
+        return totalGuests * pricePerPerson;
     }
 
-    @Override
+
     public boolean isAvailable(LocalDate startDate, LocalDate endDate, int guests) {
-        if (guests > maxCapacity) {
+        if (!startDate.isEqual(endDate)) {
             return false;
         }
 
         for (Reservation reservation : getReservations()) {
-            if (datesOverlap(reservation.getStartDate(), reservation.getEndDate(), startDate, endDate)) {
+            if (datesOverlap(reservation.getStartDate(), reservation.getEndDate(), startDate, startDate)) {
                 return false;
             }
         }
@@ -42,16 +51,13 @@ public class FarmStay extends Lodging{
 
     @Override
     public List<String> confirmAvailability(LocalDate startDate, LocalDate endDate, int adults, int children, int roomsNeeded) {
-        int totalGuests = adults + children;
-
-        if (!isAvailable(startDate, endDate, totalGuests)) {
-            System.out.println("La finca no está disponible para el rango de fechas seleccionado.");
-            return new ArrayList<>();
-        }
-
-        System.out.println("Descripción: " + getDescription());
-        return List.of("Capacidad máxima: " + getMaxCapacity() + " personas.");
+        System.out.println("Descripción: " + description);
+        System.out.println("Actividades: " + activities);
+        System.out.println("Comidas: " + meals);
+        return new ArrayList<>();
     }
+
+
 
 
     @Override
@@ -62,7 +68,7 @@ public class FarmStay extends Lodging{
         double adjustment = calculateDiscountOrIncrement(startDate, endDate);
         double totalAdjusted;
         System.out.println(this.toString());
-        System.out.println("Precio por noche: $" + pricePerNight);
+        System.out.println("Precio por persona: $" + pricePerNight);
         System.out.println("Precio base total: $" + baseTotalPrice);
 
         if(adjustment < 0){
@@ -76,19 +82,25 @@ public class FarmStay extends Lodging{
 
     @Override
     public String toString() {
-        return   name + '\n' +
+        return  name + '\n' +
                 "+------------------------------------+" + '\n' +
                 "Calificación: " + rating +'\n' +
-                "Descripción: " + description + '\n'
+                "Descripción: " + description + '\n' +
+                "Actividades: " + activities.toString() + '\n' +
+                "Meriendas Incluidas: " + meals.toString()
                 ;
     }
 
-    // Getter
-    public int getMaxCapacity() {
-        return maxCapacity;
+    // Getters
+    public List<String> getActivities() {
+        return activities;
     }
 
-    public double getPricePerNight() {
-        return pricePerNight;
+    public List<String> getMealsIncluded() {
+        return meals;
+    }
+
+    public double getPricePerPerson() {
+        return pricePerPerson;
     }
 }
