+++ src/main/java/com/bookstay/Lodging.java
@@ -1,4 +1,4 @@
-package com.bookstay.models;
+package com.bookstay;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
